+++ app/controllers/AddAnyDocumentsController.scala
@@ -18,8 +18,6 @@ package controllers
 
 import controllers.actions._
 import forms.AddAnyDocumentsFormProvider
-
-import javax.inject.Inject
 import models.{Index, LocalReferenceNumber, Mode}
 import pages.AddAnyDocumentsPage
 import play.api.i18n.{I18nSupport, MessagesApi}
@@ -28,6 +26,7 @@ import repositories.SessionRepository
 import uk.gov.hmrc.play.bootstrap.frontend.controller.FrontendBaseController
 import views.html.AddAnyDocumentsView
 
+import javax.inject.Inject
 import scala.concurrent.{ExecutionContext, Future}
 
 class AddAnyDocumentsController @Inject() (
