+++ app/viewmodels/checkAnswers/ConsignorsIdentitySummary.scala
@@ -16,7 +16,7 @@
 
 package viewmodels.checkAnswers
 
-import controllers.routes
+import controllers.consignors.{routes => consignorRoutes}
 import models.{CheckMode, Index, UserAnswers}
 import pages.ConsignorsIdentityPage
 import play.api.i18n.Messages
@@ -43,7 +43,7 @@ object ConsignorsIdentitySummary {
         actions = Seq(
           ActionItemViewModel(
             "site.change",
-            routes.ConsignorsIdentityController.onPageLoad(CheckMode, answers.lrn, index).url
+            consignorRoutes.ConsignorsIdentityController.onPageLoad(CheckMode, answers.lrn, index).url
           ).withVisuallyHiddenText(messages("consignorsIdentity.change.hidden"))
         )
       )
