+++ app/views/ConsignorEORIView.scala.html
@@ -15,6 +15,7 @@
  *@
 
 @import viewmodels.InputWidth._
+@import controllers.consignors.{routes => consignorRoutes}
 
 @this(
     layout: templates.Layout,
@@ -28,7 +29,7 @@
 
 @layout(pageTitle = title(form, messages("consignorEORI.title"))) {
 
-    @formHelper(action = routes.ConsignorEORIController.onSubmit(mode, lrn, index)) {
+    @formHelper(action = consignorRoutes.ConsignorEORIController.onSubmit(mode, lrn, index)) {
 
         @if(form.errors.nonEmpty) {
             @govukErrorSummary(ErrorSummaryViewModel(form))
