+++ app/views/ConsignorNameView.scala.html
@@ -15,6 +15,7 @@
  *@
 
 @import viewmodels.InputWidth._
+@import controllers.consignors.{routes => consignorRoutes}
 
 @this(
     layout: templates.Layout,
@@ -28,7 +29,7 @@
 
 @layout(pageTitle = title(form, messages("consignorName.title"))) {
 
-    @formHelper(action = routes.ConsignorNameController.onSubmit(mode, lrn, index)) {
+    @formHelper(action = consignorRoutes.ConsignorNameController.onSubmit(mode, lrn, index)) {
 
         @if(form.errors.nonEmpty) {
             @govukErrorSummary(ErrorSummaryViewModel(form))
