+++ app/views/ConsignorsIdentityView.scala.html
@@ -14,6 +14,8 @@
  * limitations under the License.
  *@
 
+@import controllers.consignors.{routes => consignorRoutes}
+
 @this(
     layout: templates.Layout,
     formHelper: FormWithCSRF,
@@ -26,7 +28,7 @@
 
 @layout(pageTitle = title(form, messages("consignorsIdentity.title"))) {
 
-    @formHelper(action = routes.ConsignorsIdentityController.onSubmit(mode, lrn, index), 'autoComplete -> "off") {
+    @formHelper(action = consignorRoutes.ConsignorsIdentityController.onSubmit(mode, lrn, index), 'autoComplete -> "off") {
 
         @if(form.errors.nonEmpty) {
             @govukErrorSummary(ErrorSummaryViewModel(form,  errorLinkOverrides = Map("value" -> "value_0")))
